- v3.0 - Added Register class - separating methods like find, add, remove element or macro etc. - Updated coding standard
