+++ src/Macdom/Compiler/Compiler.php
@@ -31,39 +31,9 @@ final class Compiler
 		REPLICATOR_PLACEHOLDERS_REPLACEMENTS_RE = '/\[(?<replacement>.*?)\]/';
 
 	/**
-	 * @var string
-	 */
-	public $contentType = Engine::DEFAULT_CONTENT_TYPE;
-
-	/**
-	 * @var array
-	 */
-	private $booleanAttributes = [];
-
-	/**
 	 * @var array
 	 */
 	private $closeTags = [];
-
-	/**
-	 * @var array
-	 */
-	private $elements = [];
-
-	/**
-	 * @var string
-	 */
-	private $elementsCustomCloseTags;
-
-	/**
-	 * @var string
-	 */
-	private $elementsCustomOpenTags;
-
-	/**
-	 * @var array
-	 */
-	private $macros = [];
 
 	/**
 	 * @var string
@@ -76,68 +46,14 @@ final class Compiler
 	private $processedToken;
 
 	/**
+	 * @var Register
+	 */
+	private $register;
+
+	/**
 	 * @var bool
 	 */
 	private $xmlSyntax = FALSE;
-
-
-	/**
-	 * @param string $attribute
-	 * @param string|array|NULL $contentType
-	 */
-	public function addBooleanAttribute(string $attribute, $contentType = NULL)
-	{
-		$booleanAttributes = Helpers::explodeString($attribute);
-		$contentType = $contentType ?? Engine::CONTENT_HTML;
-		$contentTypes = is_array($contentType) ? $contentType : [$contentType];
-
-		foreach ($booleanAttributes as $booleanAttribute) {
-			foreach ($contentTypes as $contentType) {
-				if ($this->findBooleanAttribute($booleanAttribute, $contentType)) {
-					continue;
-				}
-
-				$this->booleanAttributes[$contentType][] = $booleanAttribute;
-			}
-		}
-	}
-
-
-	public function addElement(string $element, array $settings = NULL)
-	{
-		$contentType = $settings && in_array(Engine::CONTENT_XML, $settings)
-			? Engine::CONTENT_XML
-			: Engine::CONTENT_HTML;
-
-		$openTags = $settings['openTags'] ?? [];
-		$closeTags = $settings['closeTags'] ?? [];
-
-		$this->addElementCustomTags($openTags, $closeTags);
-		$this->elements[$contentType][$element] = $settings ?? [];
-	}
-
-
-	public function addMacro(string $keyword, Callable $macro, array $flags = NULL)
-	{
-		$contentType = $flags && in_array(Engine::CONTENT_XML, $flags)
-			? Engine::CONTENT_XML
-			: Engine::CONTENT_HTML;
-
-		$this->macros[$contentType][$keyword] = [
-			'flags' => $flags ?? [],
-			'callback' => $macro
-		];
-	}
-
-
-	public function changeElementQuickAttributes(string $element, array $quickAttributes)
-	{
-		if ( ! $this->findElement($element)) {
-			throw new SetupException('Can\'t change quick attributes for undefined element "' . $element . '"');
-		}
-
-		$this->elements[$element]['quickAttributes'] = $quickAttributes;
-	}
 
 
 	public function compile(array $parserOutput): string
@@ -148,7 +64,7 @@ final class Compiler
 		foreach ($parserOutput['tokens'] as $token) {
 			$this->processedToken = $token;
 
-			if ( ! ! $token->text) {
+			if ($token->text) {
 				$this->addCloseTags($token->indentationLevel);
 			}
 
@@ -168,91 +84,14 @@ final class Compiler
 	}
 
 
-	public function findElement(string $element, string $contentType = NULL): bool
-	{
-		return $this->getElement($element, $contentType) !== NULL;
-	}
-
-
-	public function findMacro(string $macro, string $contentType = NULL): bool
-	{
-		return (bool) $this->getMacro($macro, $contentType);
-	}
-
-
-	/**
-	 * @param string $element
-	 * @param string|NULL $contentType
-	 * @return array|NULL
-	 */
-	public function getElement(string $element, string $contentType = NULL)
-	{
-		$contentType = $contentType ?? $this->contentType;
-
-		return array_key_exists($element, $this->elements[$contentType])
-			? $this->elements[$contentType][$element]
-			: NULL;
-	}
-
-
-	public function getElements(bool $byContentType = NULL): array
-	{
-		return $byContentType ? $this->elements[$this->contentType] : $this->elements;
-	}
-
-
-	public function getElementsBooleanAttributes(): array
-	{
-		return $this->booleanAttributes;
-	}
-
-
-	public function getElementsCustomCloseTags(): string
-	{
-		return $this->elementsCustomCloseTags ?? '';
-	}
-
-
-	public function getElementsCustomOpenTags(): string
-	{
-		return $this->elementsCustomOpenTags ?? '';
-	}
-
-
-	public function getMacros(): array
-	{
-		return $this->macros;
-	}
-
-
-	public function getTokenText(string $string): string
-	{
-		return preg_replace('/\|$/', '', trim($string), 1);
-	}
-
-
-	public function getUnpairedElements(): array
-	{
-		$unpairedElements = [];
-
-		foreach ($this->getElements(TRUE) as $element => $settings) {
-			if (in_array(Engine::UNPAIRED_ELEMENT, $settings)) {
-				$unpairedElements[] = $element;
-			}
-		}
-
-		return $unpairedElements;
-	}
-
-
 	public function refactorToken(Token $token, string $output)
 	{
 		$tokenKeyword = Helpers::getFirstWord($output);
 		$token->type = Token::TEXT;
 		$token->keyword = $tokenKeyword;
 		$token->text = $output;
-		$isElement = $this->findElement($token->keyword);
-		$isMacro = $this->findMacro($token->keyword);
+		$isElement = $this->register->findElement($token->keyword);
+		$isMacro = $this->register->findMacro($token->keyword);
 
 		if ($isElement) {
 			$token->type = Token::ELEMENT;
@@ -265,82 +104,19 @@ final class Compiler
 	}
 
 
-	public function removeBooleanAttribute(string $booleanAttribute, string $contentType = NULL): self
+	public function setRegister(Register $register): self
 	{
-		$booleanAttributes = Helpers::explodeString($booleanAttribute);
-		$contentType = $contentType ?? Engine::CONTENT_HTML;
-
-		foreach ($booleanAttributes as $booleanAttribute) {
-			$booleanAttributeKey = array_search($booleanAttribute, $this->booleanAttributes[$contentType]);
-
-			if ($booleanAttributeKey === FALSE) {
-				throw new SetupException('Can\'t remove undefined boolean attribute "' . $booleanAttribute . '"');
-			}
-
-			unset($this->booleanAttributes[$contentType][$booleanAttributeKey]);
-		}
+		$this->register = $register;
 
 		return $this;
 	}
 
 
-	public function removeElement(string $element, string $contentType = NULL): self
-	{
-		$elements = Helpers::explodeString($element);
-		$contentType = $contentType ?? Engine::CONTENT_HTML;
-
-		foreach ($elements as $element) {
-			if ( ! $this->findElement($element, $contentType)) {
-				throw new SetupException('Can\'t remove undefined element "' . $element . '"');
-			}
-
-			unset($this->elements[$contentType][$element]);
-		}
-
-		return $this;
-	}
-
-
-	public function removeMacro(string $macro, string $contentType = NULL): self
-	{
-		$macros = Helpers::explodeString($macro);
-		$contentType = $contentType ?? Engine::CONTENT_HTML;
-
-		foreach ($macros as $macro) {
-			if ( ! $this->findMacro($macro)) {
-				throw new SetupException('Can\'t remove undefined macro "' . $macro . '"');
-			}
-
-			unset($this->macros[$contentType][$macro]);
-		}
-
-		return $this;
-	}
-
-
-	public function setContentType(string $contentType): self
+	public function setXmlSyntax(string $contentType): self
 	{
 		$this->xmlSyntax = in_array($contentType, [Engine::CONTENT_XHTML, Engine::CONTENT_XML], TRUE);
 
-		if ($contentType === Engine::CONTENT_XHTML) {
-			$contentType = Engine::CONTENT_HTML;
-		}
-
-		$this->contentType = $contentType;
-
 		return $this;
-	}
-
-
-	public function unsetCodePlaceholders(array $codePlaceholders, string $string): string
-	{
-		$codePlaceholders = array_reverse($codePlaceholders);
-
-		foreach ($codePlaceholders as $codePlaceholder => $code) {
-			$string = str_replace($codePlaceholder, $code, $string);
-		}
-
-		return $string;
 	}
 
 
@@ -357,87 +133,21 @@ final class Compiler
 	}
 
 
-	private function addElementCustomTags(array $openTags, array $closeTags)
-	{
-		foreach ($openTags as $openTag) {
-			if ($this->elementsCustomOpenTags
-				&& preg_match('/' . $this->elementsCustomOpenTags . '/', $openTag)
-			) {
-				continue;
-			}
-
-			if ($this->elementsCustomOpenTags) {
-				$this->elementsCustomOpenTags .= '|';
-			}
-
-			$this->elementsCustomOpenTags .= preg_quote($openTag);
-		}
-
-		foreach ($closeTags as $closeTag) {
-			if ($this->elementsCustomCloseTags
-				&& preg_match('/' . $this->elementsCustomCloseTags . '/', $closeTag)
-			) {
-				continue;
-			}
-
-			if ($this->elementsCustomCloseTags) {
-				$this->elementsCustomCloseTags .= '|';
-			}
-
-			$this->elementsCustomCloseTags .= preg_quote($closeTag);
-		}
-
-	}
-
-
 	private function addOutput(string $code)
 	{
 		$this->output .= $code;
 	}
 
 
-	private function findBooleanAttribute(string $attribute, string $contentType = NULL): bool
+	private function getTokenText(string $string): string
 	{
-		$contentType = $contentType ?? $this->contentType;
-
-		return isset($this->booleanAttributes[$contentType])
-			&& in_array($attribute, $this->booleanAttributes[$contentType]);
-	}
-
-
-	/**
-	 * @param string $macro
-	 * @param string|NULL $contentType
-	 * @return Callable|NULL
-	 */
-	private function getMacro(string $macro, string $contentType = NULL)
-	{
-		$contentType = $contentType ?? $this->contentType;
-
-		if (array_key_exists($macro, $this->macros[$contentType])) {
-			return $this->macros[$contentType][$macro]['callback'];
-
-		} else {
-			$text = $macro;
-
-			foreach ($this->macros[$contentType] as $regularExpression => $macroObject) {
-				if ( ! in_array(Token::REGULAR_EXPRESSION_MACRO, $macroObject['flags'])) {
-					continue;
-				}
-
-				if (preg_match('/' . $regularExpression . '/', $text)) {
-					return $macroObject['callback'];
-				}
-			}
-		}
-
-		return NULL;
+		return preg_replace('/\|$/', '', trim($string), 1);
 	}
 
 
 	private function processElement(Token $token): bool
 	{
-		$element = $this->getElement($token->keyword);
+		$element = $this->register->getElement($token->keyword);
 		$tokenText = Helpers::removeFirstWord($token->text);
 		$elementCloseTags = $element['closeTags'] ?? NULL;
 		$skipAreasPlaceholders = [];
@@ -453,10 +163,7 @@ final class Compiler
 			self::BOOLEAN_ATTRIBUTES => '/\s(?<attributeValue>\S+)/'
 		];
 
-		$openTag = isset($element['openTags'])
-			? $element['openTags'][0]
-			: '<';
-
+		$openTag = isset($element['openTags']) ? $element['openTags'][0] : '<';
 		$elementToken = $openTag . $token->keyword;
 
 		foreach ($attributesTypes as $typeName => $regularExpression) {
@@ -488,7 +195,7 @@ final class Compiler
 					];
 
 				} elseif ($typeName === self::BOOLEAN_ATTRIBUTES) {
-					if ($this->findBooleanAttribute($match['attributeValue'])) {
+					if ($this->register->findBooleanAttribute($match['attributeValue'])) {
 						$elementAttributes[self::BOOLEAN_ATTRIBUTES][] = $match['attributeValue'];
 
 					} else {
@@ -507,19 +214,19 @@ final class Compiler
 			$hasHtmlClassAttribute = isset($elementAttributes[self::HTML_ATTRIBUTES]['class']);
 
 			if ($hasClassSelector || $hasHtmlClassAttribute) {
+				$classSelectors = '';
+
 				if ($hasClassSelector) {
-					$classSelectors = implode(' ', $elementAttributes[self::CLASS_SELECTORS]);
+					$classSelectors .= implode(' ', $elementAttributes[self::CLASS_SELECTORS]);
 					unset($elementAttributes[self::CLASS_SELECTORS]);
 				}
 
 				if ($hasHtmlClassAttribute) {
-					$classesFromAttribute = str_replace(["'", '"'], '',
-						$elementAttributes[self::HTML_ATTRIBUTES]['class']);
+					$classesFromAttribute = str_replace(
+						["'", '"'], '', $elementAttributes[self::HTML_ATTRIBUTES]['class']
+					);
 
-					$classSelectors = isset($classSelectors)
-						? $classSelectors . ' ' . $classesFromAttribute
-						: $classesFromAttribute;
-
+					$classSelectors .= (bool) $classSelectors ? ' ' . $classesFromAttribute : $classesFromAttribute;
 					unset($elementAttributes[self::HTML_ATTRIBUTES]['class']);
 
 					if ($attributesType === self::HTML_ATTRIBUTES) {
@@ -580,18 +287,15 @@ final class Compiler
 
 					} else {
 						throw new CompileException(
-							'Element quick attribute with key "' . $attributeKey
-							. '" doesn\'t exist on line ' . $token->line . ' near "'
-							. $tokenText . '". Quick attributes for element "'
+							'Element quick attribute with key "' . $attributeKey . '" doesn\'t exist on line '
+							. $token->line . ' near "' . $tokenText . '". Quick attributes for element "'
 							. $token->keyword . '" are "' . implode(', ', $element[self::QUICK_ATTRIBUTES]) . '"');
 					}
 				}
 
 			} elseif ($attributesType === self::BOOLEAN_ATTRIBUTES) {
 				foreach ($attributes as $attribute) {
-					$elementToken .= $this->xmlSyntax
-						? ' ' . $attribute . '="' . $attribute . '"'
-						: ' ' . $attribute;
+					$elementToken .= $this->xmlSyntax ? ' ' . $attribute . '="' . $attribute . '"' : ' ' . $attribute;
 				}
 			}
 		}
@@ -604,14 +308,12 @@ final class Compiler
 			}
 
 		} else {
-			$closeTag = $elementCloseTags
-				? $element['closeTags'][0]
-				: '</' . $token->keyword . '>';
-
+			$closeTag = $elementCloseTags ? $element['closeTags'][0] : '</' . $token->keyword . '>';
 			$closeTag = [
 				'tag' => $closeTag,
 				'indentationLevel' => $token->indentationLevel
 			];
+
 			array_unshift($this->closeTags, $closeTag);
 		}
 
@@ -638,7 +340,7 @@ final class Compiler
 
 	private function processMacro(Token $token): bool
 	{
-		$macro = $this->getMacro($token->keyword);
+		$macro = $this->register->getMacro($token->keyword);
 		$token->text = Helpers::removeFirstWord($token->text);
 		$output = $macro($this->getTokenText($token->text), $token->keyword);
 		$this->refactorToken($token, $output);
@@ -655,13 +357,12 @@ final class Compiler
 		if (preg_match_all(self::REPLICATOR_PLACEHOLDERS_REPLACEMENTS_RE, $synchronizedText, $matches)) {
 			$matches = $matches['replacement'];
 			$replicatedText = preg_replace_callback(self::REPLICATOR_PLACEHOLDERS_RE,
-				function () use (&$matches) {
+				function () use (&$matches)
+				{
 					return array_shift($matches);
 				}, $replicatedText);
 
-			$synchronizedText =
-				preg_replace(self::REPLICATOR_PLACEHOLDERS_REPLACEMENTS_RE, '', $synchronizedText);
-
+			$synchronizedText = preg_replace(self::REPLICATOR_PLACEHOLDERS_REPLACEMENTS_RE, '', $synchronizedText);
 			$replicatedText = preg_replace(self::REPLICATOR_PLACEHOLDERS_RE, '', $replicatedText);
 		}
 
@@ -678,6 +379,18 @@ final class Compiler
 		$this->addOutput($text);
 
 		return TRUE;
+	}
+
+
+	private function unsetCodePlaceholders(array $codePlaceholders, string $string): string
+	{
+		$codePlaceholders = array_reverse($codePlaceholders);
+
+		foreach ($codePlaceholders as $codePlaceholder => $code) {
+			$string = str_replace($codePlaceholder, $code, $string);
+		}
+
+		return $string;
 	}
 
 }
